requirements.txt
flask
flask-cors
opencv-python
numpy
face_recognition
scikit-image
torch
ultralytics
pyjwt
werkzeug
pillow
requests
